movgrab-1.1.14

+++ libUseful-2.0/sound.c
@@ -163,7 +163,7 @@ switch(Header.Format)
 		break;
 }
 
-STREAMSeek(S,Header.DataStart,SEEK_SET);
+STREAMSeek(S,(double) Header.DataStart,SEEK_SET);
 return(AudioInfo);
 }
